services/telegram-manager/requirements.txt
# FILE: ./services/telegram-manager/requirements.txt
python-telegram-bot==13.15
requests
python-dotenv
python-json-logger
schedule